Clean up node compare to use Comparer directly.
