+++ src/Hyperbee.Json/Descriptors/Element/Functions/LengthElementFunction.cs
@@ -1,6 +1,8 @@
 ﻿using System.Linq.Expressions;
 using System.Text.Json;
+using Hyperbee.Json.Descriptors.Types;
 using Hyperbee.Json.Filters.Parser;
+using ValueType = Hyperbee.Json.Descriptors.Types.ValueType;
 
 namespace Hyperbee.Json.Descriptors.Element.Functions;
