+++ src/Hyperbee.Json/Descriptors/Element/Functions/SearchElementFunction.cs
@@ -1,8 +1,10 @@
 ﻿using System.Linq.Expressions;
 using System.Text.Json;
 using System.Text.RegularExpressions;
+using Hyperbee.Json.Descriptors.Types;
 using Hyperbee.Json.Filters;
 using Hyperbee.Json.Filters.Parser;
+using ValueType = Hyperbee.Json.Descriptors.Types.ValueType;
 
 namespace Hyperbee.Json.Descriptors.Element.Functions;
