+++ src/Hyperbee.Json/Descriptors/ITypeDescriptor.cs
@@ -1,4 +1,5 @@
-﻿using Hyperbee.Json.Filters;
+﻿using Hyperbee.Json.Descriptors.Types;
+using Hyperbee.Json.Filters;
 using Hyperbee.Json.Filters.Parser;
 
 namespace Hyperbee.Json.Descriptors;
