+++ src/Hyperbee.Json/Descriptors/Node/Functions/CountNodeFunction.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq.Expressions;
 using System.Text.Json.Nodes;
+using Hyperbee.Json.Descriptors.Types;
 using Hyperbee.Json.Filters.Parser;
 
 namespace Hyperbee.Json.Descriptors.Node.Functions;
