+++ src/Hyperbee.Json/Descriptors/Node/Functions/SearchNodeFunction.cs
@@ -2,8 +2,10 @@
 using System.Text.Json;
 using System.Text.Json.Nodes;
 using System.Text.RegularExpressions;
+using Hyperbee.Json.Descriptors.Types;
 using Hyperbee.Json.Filters;
 using Hyperbee.Json.Filters.Parser;
+using ValueType = Hyperbee.Json.Descriptors.Types.ValueType;
 
 namespace Hyperbee.Json.Descriptors.Node.Functions;
