+++ src/Hyperbee.Json/Descriptors/Node/NodeTypeDescriptor.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text.Json.Nodes;
 using Hyperbee.Json.Descriptors.Node.Functions;
+using Hyperbee.Json.Descriptors.Types;
 using Hyperbee.Json.Filters;
 
 namespace Hyperbee.Json.Descriptors.Node;
