+++ src/Hyperbee.Json/Descriptors/Types/INodeType.cs
@@ -1,0 +1,8 @@
+﻿namespace Hyperbee.Json.Descriptors.Types;
+
+public interface INodeType
+{
+    public NodeTypeKind Kind { get; }
+
+    public INodeTypeComparer Comparer { get; set; }
+}
