+++ src/Hyperbee.Json/Descriptors/Types/NodeTypeKind.cs
@@ -1,0 +1,10 @@
+﻿namespace Hyperbee.Json.Descriptors.Types;
+
+public enum NodeTypeKind
+{
+    Null,
+    Nothing,
+    Value,
+    Node,
+    NodeList
+}
