+++ src/Hyperbee.Json/Descriptors/Types/Nothing.cs
@@ -1,0 +1,7 @@
+﻿namespace Hyperbee.Json.Descriptors.Types;
+
+public struct Nothing : INodeType
+{
+    public NodeTypeKind Kind => NodeTypeKind.Nothing;
+    public INodeTypeComparer Comparer { get; set; }
+}
