+++ src/Hyperbee.Json/Descriptors/Types/Null.cs
@@ -1,0 +1,7 @@
+﻿namespace Hyperbee.Json.Descriptors.Types;
+
+public struct Null : INodeType
+{
+    public NodeTypeKind Kind => NodeTypeKind.Null;
+    public INodeTypeComparer Comparer { get; set; }
+}
