+++ src/Hyperbee.Json/Filters/Parser/Expressions/JsonExpressionFactory.cs
@@ -1,5 +1,5 @@
 ﻿using System.Linq.Expressions;
-using Hyperbee.Json.Descriptors;
+using Hyperbee.Json.Descriptors.Types;
 
 namespace Hyperbee.Json.Filters.Parser.Expressions;
