+++ src/Hyperbee.Json/Filters/Parser/Expressions/LiteralExpressionFactory.cs
@@ -1,6 +1,6 @@
 ﻿using System.Linq.Expressions;
-using Hyperbee.Json.Descriptors;
-using ValueType = Hyperbee.Json.Descriptors.ValueType;
+using Hyperbee.Json.Descriptors.Types;
+using ValueType = Hyperbee.Json.Descriptors.Types.ValueType;
 
 namespace Hyperbee.Json.Filters.Parser.Expressions;
