+++ src/Hyperbee.Json/Filters/Parser/NodeTypeExpression.cs
@@ -1,33 +1,13 @@
 ﻿using System.Linq.Expressions;
 using System.Reflection;
-using Hyperbee.Json.Descriptors;
+using Hyperbee.Json.Descriptors.Types;
 
 namespace Hyperbee.Json.Filters.Parser;
 
-public static class NodeTypeComparerBinderExpression<TNode>
-{
-    private static readonly Expression BindComparerExpressionConst = Expression.Constant( (Func<FilterParserContext<TNode>, INodeType, INodeType>) BindComparer );
-    internal static Expression BindComparerExpression( FilterParserContext<TNode> parserContext, Expression expression )
-    {
-        if ( expression == null )
-            return null;
-
-        var parserContextExp = Expression.Constant( parserContext );
-
-        return Expression.Invoke( BindComparerExpressionConst, parserContextExp,
-            Expression.Convert( expression, typeof( INodeType ) ) );
-    }
-
-    internal static INodeType BindComparer( FilterParserContext<TNode> parserContext, INodeType item )
-    {
-        item.SetComparer( parserContext.Descriptor.Comparer );
-        return item;
-    }
-
-}
-
 public static class NodeTypeExpression<TNode>
 {
+    // TODO: Add, Add, Or, and Not operators
+
     private static readonly MethodInfo AreEqualMethodInfo = typeof( NodeTypeExpression<TNode> ).GetMethod( nameof( AreEqual ) );
     private static readonly MethodInfo AreNotEqualMethodInfo = typeof( NodeTypeExpression<TNode> ).GetMethod( nameof( AreNotEqual ) );
     private static readonly MethodInfo IsLessThanMethodInfo = typeof( NodeTypeExpression<TNode> ).GetMethod( nameof( IsLessThan ) );
@@ -43,110 +23,10 @@ public static class NodeTypeExpression<TNode>
     public static Expression GreaterThanOrEqual( Expression left, Expression right ) => Expression.Call( IsGreaterThanOrEqualMethodInfo, left, right );
 
 
-    public static bool AreEqual( INodeType left, INodeType right )
-    {
-        return left switch
-        {
-            // Symmetric compares
-            ValueType<string> leftNode when right is ValueType<string> rightNode => leftNode.Value == rightNode.Value,
-            ValueType<bool> leftNode when right is ValueType<bool> rightNode => leftNode.Value == rightNode.Value,
-            ValueType<float> leftNode when right is ValueType<float> rightNode => leftNode == rightNode, // TODO: should be pass through for tolerance check
-            NodesType<TNode> leftNode when right is NodesType<TNode> rightNode => leftNode == rightNode,
-            Nothing leftNode when right is Nothing rightNode => true,
-            Null leftNode when right is Null rightNode => true,
-
-            // Asymmetric compares
-            ValueType<string> leftNode when right is NodesType<TNode> rightNode => leftNode == rightNode,
-            ValueType<bool> leftNode when right is NodesType<TNode> rightNode => leftNode == rightNode,
-            ValueType<float> leftNode when right is NodesType<TNode> rightNode => leftNode == rightNode,
-            NodesType<TNode> leftNode => leftNode == right,
-
-            _ => false
-        };
-    }
-
-    public static bool AreNotEqual( INodeType left, INodeType right )
-    {
-        return !AreEqual( left, right );
-    }
-
-    public static bool IsLessThan( INodeType left, INodeType right )
-    {
-        return left switch
-        {
-            // Direct compare
-            ValueType<string> leftNode when right is ValueType<string> rightNode => leftNode.Value == rightNode.Value,
-            ValueType<bool> leftNode when right is ValueType<bool> rightNode => leftNode.Value == rightNode.Value,
-            ValueType<float> leftNode when right is ValueType<float> rightNode => leftNode < rightNode, // TODO: should be pass through for tolerance check
-            NodesType<TNode> leftNode when right is NodesType<TNode> rightNode => leftNode < rightNode,
-
-            // Asymmetric compares
-            ValueType<string> leftNode when right is NodesType<TNode> rightNode => leftNode < rightNode,
-            ValueType<bool> leftNode when right is NodesType<TNode> rightNode => leftNode < rightNode,
-            ValueType<float> leftNode when right is NodesType<TNode> rightNode => leftNode < rightNode,
-            NodesType<TNode> leftNode => leftNode < right,
-
-            _ => false
-        };
-    }
-
-    public static bool IsGreaterThan( INodeType left, INodeType right )
-    {
-        return left switch
-        {
-            // Direct compare
-            ValueType<string> leftNode when right is ValueType<string> rightNode => leftNode.Value == rightNode.Value,
-            ValueType<bool> leftNode when right is ValueType<bool> rightNode => leftNode.Value == rightNode.Value,
-            ValueType<float> leftNode when right is ValueType<float> rightNode => leftNode > rightNode, // TODO: should be pass through for tolerance check
-            NodesType<TNode> leftNode when right is NodesType<TNode> rightNode => leftNode > rightNode,
-
-            // Asymmetric compares
-            ValueType<string> leftNode when right is NodesType<TNode> rightNode => leftNode > rightNode,
-            ValueType<bool> leftNode when right is NodesType<TNode> rightNode => leftNode > rightNode,
-            ValueType<float> leftNode when right is NodesType<TNode> rightNode => leftNode > rightNode,
-            NodesType<TNode> leftNode => leftNode > right,
-
-            _ => false
-        };
-    }
-
-    public static bool IsLessThanOrEqual( INodeType left, INodeType right )
-    {
-        return left switch
-        {
-            // Direct compare
-            ValueType<string> leftNode when right is ValueType<string> rightNode => leftNode.Value == rightNode.Value,
-            ValueType<bool> leftNode when right is ValueType<bool> rightNode => leftNode.Value == rightNode.Value,
-            ValueType<float> leftNode when right is ValueType<float> rightNode => leftNode <= rightNode, // TODO: should be pass through for tolerance check
-            NodesType<TNode> leftNode when right is NodesType<TNode> rightNode => leftNode <= rightNode,
-
-            // Asymmetric compares
-            ValueType<string> leftNode when right is NodesType<TNode> rightNode => leftNode <= rightNode,
-            ValueType<bool> leftNode when right is NodesType<TNode> rightNode => leftNode <= rightNode,
-            ValueType<float> leftNode when right is NodesType<TNode> rightNode => leftNode <= rightNode,
-            NodesType<TNode> leftNode => leftNode <= right,
-
-            _ => false
-        };
-    }
-
-    public static bool IsGreaterThanOrEqual( INodeType left, INodeType right )
-    {
-        return left switch
-        {
-            // Direct compare
-            ValueType<string> leftNode when right is ValueType<string> rightNode => leftNode.Value == rightNode.Value,
-            ValueType<bool> leftNode when right is ValueType<bool> rightNode => leftNode.Value == rightNode.Value,
-            ValueType<float> leftNode when right is ValueType<float> rightNode => leftNode >= rightNode, // TODO: should be pass through for tolerance check
-            NodesType<TNode> leftNode when right is NodesType<TNode> rightNode => leftNode >= rightNode,
-
-            // Asymmetric compares
-            ValueType<string> leftNode when right is NodesType<TNode> rightNode => leftNode >= rightNode,
-            ValueType<bool> leftNode when right is NodesType<TNode> rightNode => leftNode >= rightNode,
-            ValueType<float> leftNode when right is NodesType<TNode> rightNode => leftNode >= rightNode,
-            NodesType<TNode> leftNode => leftNode >= right,
-
-            _ => false
-        };
-    }
+    public static bool AreEqual( INodeType left, INodeType right ) => left.Comparer.Compare( left, right, Operator.Equals ) == 0;
+    public static bool AreNotEqual( INodeType left, INodeType right ) => left.Comparer.Compare( left, right, Operator.NotEquals ) != 0;
+    public static bool IsLessThan( INodeType left, INodeType right ) => left.Comparer.Compare( left, right, Operator.LessThan ) < 0;
+    public static bool IsLessThanOrEqual( INodeType left, INodeType right ) => left.Comparer.Compare( left, right, Operator.LessThanOrEqual ) <= 0;
+    public static bool IsGreaterThan( INodeType left, INodeType right ) => left.Comparer.Compare( left, right, Operator.GreaterThan ) > 0;
+    public static bool IsGreaterThanOrEqual( INodeType left, INodeType right ) => left.Comparer.Compare( left, right, Operator.GreaterThanOrEqual ) >= 0;
 }
